+++ packages/runtime/adapters/eventstore-adapters/eventstore-base/src/types.ts
@@ -308,12 +308,14 @@ export type ExportOptions = {
   cursor: InputCursor
   maintenanceMode: MAINTENANCE_MODE
   bufferSize: number
+  preferRegularEventLoader: boolean
 }
 
 export type ExportEventsStream = stream.Readable & {
   readonly cursor: InputCursor
   readonly isBufferOverflow: boolean
   readonly isEnd: boolean
+  readonly preferRegularEventLoader: boolean
 }
 
 export type ImportSecretsOptions = {
@@ -452,6 +454,12 @@ export interface AdapterFunctions<ConfiguredProps extends {}> {
     ConfiguredProps,
     NonNullable<AdapterPoolBoundProps['getEventLoaderNative']>
   >
+
+  runtimeInfo: PoolMethod<ConfiguredProps, Adapter['runtimeInfo']>
+  setReconnectionMode?: PoolMethod<
+    ConfiguredProps,
+    Adapter['setReconnectionMode']
+  >
 }
 
 export interface EventLoader {
@@ -463,6 +471,17 @@ export interface EventLoader {
 
 export type EventLoaderOptions = {
   preferRegular: boolean // prefer regular implementation via loadEvents over native one
+}
+
+export type AdapterRuntimeInfo = {
+  connectionCount: number
+  disposed: boolean
+  [key: string]: any
+}
+
+export type ReconnectionMode = {
+  maxReconnectionTimes?: number
+  delayBeforeReconnection?: number
 }
 
 export interface Adapter extends CoreEventstore {
@@ -504,4 +523,7 @@ export interface Adapter extends CoreEventstore {
     filter: EventLoaderFilter,
     options?: EventLoaderOptions
   ) => Promise<EventLoader>
+
+  runtimeInfo: () => AdapterRuntimeInfo
+  setReconnectionMode: (mode: ReconnectionMode) => void
 }